modification

+++ apis/courses/views.py
@@ -1,4 +1,4 @@
-from flask_restx import Resource, fields, Namespace
+from flask_restx import Resource, fields, Namespace, abort
 from extensions import db
 from ..models import Student, Courses, CourseRegistered
 from http import HTTPStatus
@@ -10,9 +10,9 @@ courses_namespace = Namespace('Courses', description='Courses related operations
 
 course = courses_namespace.model('Course', {
     'id': fields.Integer(required=True, description='The course identifier'),
-    "course code": fields.String(required=True, description='The course code'),
-    "course title": fields.String(required=True, description='The course title'),
-    "course unit": fields.Integer(required=True, description='The course unit'),
+    "course_code": fields.String(required=True, description='The course code'),
+    "course_title": fields.String(required=True, description='The course title'),
+    "course_unit": fields.Integer(required=True, description='The course unit'),
     "lecturer": fields.String(required=True, description='The course lecturer')
 })
 
@@ -26,13 +26,12 @@ class Register(Resource):
     @courses_namespace.doc('register_course')
     @courses_namespace.expect(course)
     @courses_namespace.marshal_with(course, code=201)
-    @jwt_required
     def post(self):
         data = request.get_json()
         new_course = Courses(
-            course_code=data['course code'],
-            course_title=data['course title'],
-            course_unit=data['course unit'],
+            course_code=data['course_code'],
+            course_title=data['course_title'],
+            course_unit=data['course_unit'],
             lecturer=data['lecturer'],
         )
         db.session.add(new_course)
@@ -55,25 +54,29 @@ class Register(Resource):
         return course, HTTPStatus.OK
     
     @courses_namespace.doc('delete_course_by_id')
-    @courses_namespace.marshal_with(course, code=200)
-    @jwt_required
+    # @courses_namespace.marshal_with(course, code=200)
+    
     def delete(self, course_id):
         course = Courses.query.filter_by(id=course_id).first()
+        if not course:
+            abort(404, message="Course not found")
         db.session.delete(course)
         db.session.commit()
 
-        return course, HTTPStatus.OK
+        return {"message": "course deleted"}, HTTPStatus.OK
     
     @courses_namespace.doc('update_course_by_id')
     @courses_namespace.expect(course)
     @courses_namespace.marshal_with(course, code=200)
-    @jwt_required
+    
     def put(self, course_id):
         course = Courses.query.filter_by(id=course_id).first()
         data = request.get_json()
-        course.course_code = data['course code']
-        course.course_title = data['course title']
-        course.course_unit = data['course unit']
+        if not course:
+            abort(404, message="Course not found")
+        course.course_code = data['course_code']
+        course.course_title = data['course_title']
+        course.course_unit = data['course_unit']
         course.lecturer = data['lecturer']
         db.session.commit()
 
@@ -104,7 +107,7 @@ class Register(Resource):
     
     @courses_namespace.doc('delete_course_registered_by_id')
     @courses_namespace.marshal_with(courses_registered, code=200)
-    @jwt_required
+    
     def delete(self, course_id):
         courses_registered = CourseRegistered.query.filter_by(course_id=course_id).all()
         for course_registered in courses_registered:
